Fixed wrong column type in migration

+++ database/migrations/2018_07_30_090338_add_category_id_to_posts_table.php
@@ -14,7 +14,7 @@ class AddCategoryIdToPostsTable extends Migration
     public function up()
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->tinyInteger('category_id')->nullable()->after('brand_id');
+            $table->integer('category_id')->nullable()->after('brand_id');
         });
     }
